Eslint/Prettier configuration with git hooks
