+++ .gitignore
@@ -1,2 +1,4 @@
 # npm
 node_modules
+
+.idea
