+++ .prettierrc
@@ -1,0 +1,9 @@
+{
+  "printWidth": 120,
+  "arrowParens": "always",
+  "semi": true,
+  "tabWidth": 4,
+  "jsxSingleQuote": true,
+  "singleQuote": true,
+  "trailingComma": "es5"
+}
