--- index.js
@@ -1,12 +1,0 @@
-const { ApolloServer} = require('apollo-server')
-const { typeDefs } = require('./graphql/typeDefs')
-const { resolvers } = require('./graphql/resolvers')
-
-const server = new ApolloServer({
-  typeDefs,
-  resolvers,
-})
-
-server.listen().then(({url}) => {
-  console.log(`Server ready at ${url}`)
-})
