+++ package.json
@@ -1,19 +1,37 @@
 {
-  "name": "line-backend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "apollo-server": "^2.19.2",
-    "graphql": "^15.4.0",
-    "nodemon": "^2.0.7",
-    "uuid": "^8.3.2"
-  }
+    "name": "line-backend",
+    "version": "1.0.0",
+    "description": "",
+    "main": "src/index.js",
+    "scripts": {
+        "start": "node src/index.js",
+        "dev": "nodemon src/index.js",
+        "lint": "lint-staged",
+        "postinstall": "husky install",
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "apollo-server": "^2.19.2",
+        "graphql": "^15.4.0",
+        "nodemon": "^2.0.7",
+        "uuid": "^8.3.2"
+    },
+    "devDependencies": {
+        "eslint": "^7.20.0",
+        "eslint-config-prettier": "^7.2.0",
+        "eslint-plugin-prettier": "^3.3.1",
+        "husky": "^5.0.9",
+        "lint-staged": "^10.5.4",
+        "prettier": "^2.2.1"
+    },
+    "lint-staged": {
+        "*.+(js)": [
+            "eslint src --fix"
+        ],
+        "*.+(json|md)": [
+            "prettier --write"
+        ]
+    }
 }
