+++ src/index.js
@@ -1,0 +1,15 @@
+const { ApolloServer } = require('apollo-server');
+const { typeDefs } = require('./graphql/typeDefs');
+const { resolvers } = require('./graphql/resolvers');
+
+const server = new ApolloServer({
+    typeDefs,
+    resolvers,
+});
+
+const s =
+    'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa';
+
+server.listen().then(({ url }) => {
+    console.log(`Server ready at ${url}`);
+});
